COMMUNICATION-1614 - New protocol development Schneider PM5561 ModBus meter

+++ upl/com.energyict.mdc.upl.impl/protocolimpl/src/main/java/com/energyict/protocolimpl/modbus/core/AbstractRegister.java
@@ -121,7 +121,7 @@ public class AbstractRegister {
 		}
     }
 
-    public ReadGeneralReferenceRequest getReadGeneralReferenceRequest(long referenceSpec) throws IOException {
+    public ReadGeneralReferenceRequest getReadGeneralReferenceRequest(int referenceSpec) throws IOException {
         try {
             final ReadGeneralReferenceRequest request = getRegisterFactory().getFunctionCodeFactory().readGeneralReferenceRequest(referenceSpec);
             return request;
